Implemented ball, and paddle as separate classes. Controls implemented (Minimal). Collision detection implemented(basic).
Improved Brick generation. Switched from 2D array to list.
Collision works but colliding with the sides of the bricks is a buggy.
Paddle movement implemented but no collision implemented for it.
When ball collides with the right and bottom edges of the canvas the ball is actually outside. To Fix.
Brick break on hit. Minimal tests ran on thoughness and color changing.